Rendering now working
A triangle is now rendered. Now calling RenderWrapper instead of Render in
Game, will now clear and flip the back buffer.

Minor but essential modifications to resize routine, resetting the render target to the new back buffer pointer.

Minor changes to remove warnings (casts, mostly from size_t to UINT, ugh).

+++ Code/Framework/Game.hpp
@@ -50,7 +50,7 @@ namespace Framework
 			rendered when this is called. Should be used to render everything
 			else.
 		*/
-		virtual void Render(double dt, float interpolation) = 0;
+		virtual void Render(double dt, double interpolation) = 0;
 
 		/**
 			Get a reference to the window.
@@ -74,7 +74,7 @@ namespace Framework
 			Render a scene. Will render the entities in the scene manager,
 			and then give control to the derived class by calling Render().
 		*/
-		void RenderWrapper(float dt, float interpolation);
+		void RenderWrapper(double dt, double interpolation);
 	};
 }
